refactor: move HotkeyLabel & ContextMenu to atoms
