+++ src/components/atoms/HotkeyLabel/HotkeyLabel.styled.tsx
