+++ src/components/atoms/HotkeyLabel/HotkeyLabel.tsx
