+++ src/components/atoms/index.ts
@@ -1,6 +1,8 @@
 export {default as CopyButton} from './CopyButton';
+export {default as ContextMenu} from './ContextMenu';
 export {default as Dots} from './Dots';
 export {default as FileExplorer} from './FileExplorer';
+export {default as HotkeyLabel} from './HotkeyLabel';
 export {default as Icon} from './Icon';
 export {default as ImageIcon} from './ImageIcon';
 export {default as KeyValueInput} from './KeyValueInput';
