+++ src/navsections/HelmChartSectionBlueprint/HelmChartContextMenu.tsx
@@ -18,9 +18,7 @@ import {isInPreviewModeSelector} from '@redux/selectors';
 import {getAbsoluteFilePath} from '@redux/services/fileEntry';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
 
-import {ContextMenu} from '@molecules';
-
-import {Dots} from '@atoms';
+import {ContextMenu, Dots} from '@atoms';
 
 import {useCreate, useDuplicate, useFilterByFileOrFolder, useProcessing, useRename} from '@hooks/fileTreeHooks';
