+++ src/navsections/K8sResourceSectionBlueprint/ResourceKindContextMenuWrapper.tsx
@@ -24,7 +24,7 @@ import {getResourcesForPath} from '@redux/services/fileEntry';
 import {isFileResource, isUnsavedResource} from '@redux/services/resource';
 import {removeResources} from '@redux/thunks/removeResources';
 
-import {ContextMenu} from '@molecules';
+import {ContextMenu} from '@atoms';
 
 function deleteResourceWithConfirm(resource: K8sResource, resourceMap: ResourceMapType, dispatch: AppDispatch) {
   let title = `Are you sure you want to delete ${resource.name}?`;
